1. Code cleanup 2. Edit test UI improvement

+++ app/controllers/admin/HomeController.java
@@ -75,6 +75,11 @@ public class HomeController extends Controller {
     @ValidationAction.ValidationActivity(validationActionType = Test.class)
     public Result editTest(){
         Form<Test> testForm = formFactory.form(Test.class).bindFromRequest();
+        testForm.get().testQuestions.forEach(testQuestion -> testQuestion.testAnswers.forEach(testAnswer -> {
+            if(testAnswer.isCorrect == null){
+                testAnswer.isCorrect = false;
+            }
+        }));
         testForm.get().update();
         return redirect("/");
     }
@@ -82,16 +87,20 @@ public class HomeController extends Controller {
     @AdminOnly
     public Result activateTest(Long testId){
         Test test = Test.find.byId(testId);
-        test.testStatus = Test.TestStatus.ACTIVE;
-        test.update();
+        if (test != null) {
+            test.testStatus = Test.TestStatus.ACTIVE;
+            test.update();
+        }
         return ok();
     }
 
     @AdminOnly
     public Result deactivateTest(Long testId){
         Test test = Test.find.byId(testId);
-        test.testStatus = Test.TestStatus.DRAFT;
-        test.update();
+        if (test != null) {
+            test.testStatus = Test.TestStatus.DRAFT;
+            test.update();
+        }
         return ok();
     }
 
@@ -104,7 +113,9 @@ public class HomeController extends Controller {
     @AdminOnly
     public Result deleteTest(Long testId){
         Test test = Test.find.byId(testId);
-        test.delete();
+        if (test != null) {
+            test.delete();
+        }
         return ok();
     }
